+++ src/main/java/roomescape/controller/dto/ReservationRequest.java
@@ -1,6 +1,6 @@
 package roomescape.controller.dto;
 
-import roomescape.service.dto.ReservationInput;
+import roomescape.service.dto.input.ReservationInput;
 
 public record ReservationRequest(String name, String date, Long timeId) {
