+++ src/main/java/roomescape/controller/dto/ReservationTimeRequest.java
@@ -1,6 +1,6 @@
 package roomescape.controller.dto;
 
-import roomescape.service.dto.ReservationTimeInput;
+import roomescape.service.dto.input.ReservationTimeInput;
 
 public record ReservationTimeRequest(String startAt) {
