+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -8,8 +8,8 @@ import roomescape.domain.ReservationTime;
 import roomescape.exception.ExistReservationInReservationTimeException;
 import roomescape.exception.NotExistReservationTimeException;
 import roomescape.exception.ReservationTimeAlreadyExistsException;
-import roomescape.service.dto.ReservationTimeInput;
-import roomescape.service.dto.ReservationTimeOutput;
+import roomescape.service.dto.input.ReservationTimeInput;
+import roomescape.service.dto.output.ReservationTimeOutput;
 
 @Service
 public class ReservationTimeService {
