+++ src/main/java/roomescape/service/dto/input/ReservationInput.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.input;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
