+++ src/main/java/roomescape/service/dto/input/ReservationTimeInput.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.input;
 
 import roomescape.domain.ReservationTime;
