+++ src/main/java/roomescape/service/dto/output/ReservationOutput.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.output;
 
 import java.util.List;
 import roomescape.domain.Reservation;
